fixed review comments

+++ pkg/driver/controllerserver.go
@@ -51,6 +51,7 @@ func (cs *controllerServer) CreateVolume(_ context.Context, req *csi.CreateVolum
 		pvcName            string
 		pvcNamespace       string
 		bucketVersioning   string
+		quotaLimitEnabled  bool
 	)
 
 	modifiedRequest, err := utils.ReplaceAndReturnCopy(req)
@@ -61,7 +62,7 @@ func (cs *controllerServer) CreateVolume(_ context.Context, req *csi.CreateVolum
 
 	volumeName, err := sanitizeVolumeID(req.GetName())
 	if err != nil {
-		return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("Error in sanitizeVolumeID  %v", err))
+		return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("Error in sanitizeVolumeID %v", err))
 	}
 	volumeID := volumeName
 	if len(volumeID) == 0 {
@@ -233,26 +234,28 @@ func (cs *controllerServer) CreateVolume(_ context.Context, req *csi.CreateVolum
 			klog.Infof("Created bucket: %s", bucketName)
 		}
 
-		quotaBytes := req.GetCapacityRange().GetRequiredBytes()
-		if quotaBytes <= 0 {
-			if params["userProvidedBucket"] == "false" {
-				_ = sess.DeleteBucket(bucketName) // #nosec G104
+		if quotaLimitEnabled {
+			quotaBytes := req.GetCapacityRange().GetRequiredBytes()
+			if quotaBytes <= 0 {
+				if params["userProvidedBucket"] == "false" {
+					_ = sess.DeleteBucket(bucketName) // #nosec G104 -- intentional: best-effort cleanup in error path, don't mask original error
+				}
+				return nil, status.Error(codes.InvalidArgument, "quotaLimit enabled but no positive storage size requested in PVC")
 			}
-			return nil, status.Error(codes.InvalidArgument, "quotaLimit enabled but no positive storage size requested in PVC")
-		}
 
-		resConfApikey := secretMap[constants.ResConfApiKey]
+			resConfApikey := secretMap[constants.ResConfApiKey]
 
-		klog.Infof("Applying hard quota of %d bytes to bucket %s", quotaBytes, bucketName)
-		err = s3client.UpdateQuotaLimit(quotaBytes, resConfApikey, bucketName, endPoint, creds.IAMEndpoint)
-		if err != nil {
-			klog.Errorf("Failed to set quota limit on bucket %s: %v", bucketName, err)
-			if params["userProvidedBucket"] == "false" {
-				_ = sess.DeleteBucket(bucketName)
+			klog.Infof("Applying hard quota of %d bytes to bucket %s", quotaBytes, bucketName)
+			err = s3client.UpdateQuotaLimit(quotaBytes, resConfApikey, bucketName, endPoint, creds.IAMEndpoint)
+			if err != nil {
+				klog.Errorf("Failed to set quota limit on bucket %s: %v", bucketName, err)
+				if params["userProvidedBucket"] == "false" {
+					_ = sess.DeleteBucket(bucketName)
+				}
+				return nil, status.Error(codes.Internal, fmt.Sprintf("failed to set bucket quota limit: %v", err))
 			}
-			return nil, status.Error(codes.Internal, fmt.Sprintf("failed to set bucket quota limit: %v", err))
+			klog.Infof("Successfully applied hard quota %d bytes to bucket %s", quotaBytes, bucketName)
 		}
-		klog.Infof("Successfully applied hard quota %d bytes to bucket %s", quotaBytes, bucketName)
 
 		if bucketVersioning != "" {
 			enable := strings.ToLower(strings.TrimSpace(bucketVersioning)) == "true"
@@ -285,22 +288,24 @@ func (cs *controllerServer) CreateVolume(_ context.Context, req *csi.CreateVolum
 			return nil, status.Error(codes.PermissionDenied, fmt.Sprintf("%v: %v", err, tempBucketName))
 		}
 
-		quotaBytes := req.GetCapacityRange().GetRequiredBytes()
-		if quotaBytes <= 0 {
-			_ = sess.DeleteBucket(tempBucketName) // #nosec G104 -- intentional: best-effort cleanup in error path, don't mask original error
-			return nil, status.Error(codes.InvalidArgument, "quotaLimit enabled but no positive storage size requested")
-		}
+		if quotaLimitEnabled {
+			quotaBytes := req.GetCapacityRange().GetRequiredBytes()
+			if quotaBytes <= 0 {
+				_ = sess.DeleteBucket(tempBucketName) // #nosec G104 -- intentional: best-effort cleanup in error path, don't mask original error
+				return nil, status.Error(codes.InvalidArgument, "quotaLimit enabled but no positive storage size requested")
+			}
 
-		resConfApikey := secretMap[constants.ResConfApiKey]
+			resConfApikey := secretMap[constants.ResConfApiKey]
 
-		klog.Infof("Applying hard quota of %d bytes to temp bucket %s", quotaBytes, tempBucketName)
-		err = s3client.UpdateQuotaLimit(quotaBytes, resConfApikey, tempBucketName, endPoint, creds.IAMEndpoint)
-		if err != nil {
-			klog.Errorf("Failed to set quota limit on temp bucket %s: %v", tempBucketName, err)
-			_ = sess.DeleteBucket(tempBucketName)
-			return nil, status.Error(codes.Internal, fmt.Sprintf("failed to set bucket quota limit: %v", err))
+			klog.Infof("Applying hard quota of %d bytes to temp bucket %s", quotaBytes, tempBucketName)
+			err = s3client.UpdateQuotaLimit(quotaBytes, resConfApikey, tempBucketName, endPoint, creds.IAMEndpoint)
+			if err != nil {
+				klog.Errorf("Failed to set quota limit on temp bucket %s: %v", tempBucketName, err)
+				_ = sess.DeleteBucket(tempBucketName)
+				return nil, status.Error(codes.Internal, fmt.Sprintf("failed to set bucket quota limit: %v", err))
+			}
+			klog.Infof("Successfully applied hard quota %d bytes to temp bucket %s", quotaBytes, tempBucketName)
 		}
-		klog.Infof("Successfully applied hard quota %d bytes to temp bucket %s", quotaBytes, tempBucketName)
 
 		if bucketVersioning != "" {
 			enable := strings.ToLower(strings.TrimSpace(bucketVersioning)) == "true"
